// NullRefAcademyFront/src/app/models/studentcourses.model.ts
export interface StudentCourses{
    id: number;
    title: string;
    description: string;
    points: number;
    createdDate: string;
    categoryID: number;
    teacherID: number;
    adminID: number;
    studentsIDs: string;
    isFinished: string;
    startDate: string;
    endDate: string;
}